# frontend/src/lib/api/schema/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlametteMoney.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/transactions/{id}": {
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete transaction",
        "description": "Delete a transaction and revert account balance changes.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get transaction",
        "description": "Get a transaction by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Transactions"
        ],
        "summary": "Update transaction",
        "description": "Update a transaction and adjust account balances.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetTransactionResponse": {
        "required": [
          "id",
          "date",
          "type",
          "amount",
          "amount2",
          "currency",
          "currency2",
          "accountId",
          "tripId",
          "categoryId",
          "subCategoryId",
          "targetAccountId",
          "originalTransactionId",
          "isRefund",
          "note",
          "merchantName",
          "location",
          "items"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "amount2": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency2": {
            "type": [
              "null",
              "string"
            ]
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subCategoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetAccountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "originalTransactionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isRefund": {
            "type": "boolean"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "merchantName": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionItemResponse"
            }
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "TransactionType": {
        "enum": [
          "Income",
          "Expense",
          "Transfer",
          "Refund"
        ]
      },
...
      "UpdateTransactionRequest": {
        "required": [
          "date",
          "type",
          "amount",
          "accountId",
          "tripId",
          "categoryId",
          "subCategoryId",
          "targetAccountId",
          "originalTransactionId",
          "note",
          "merchantName",
          "location"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subCategoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetAccountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "originalTransactionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "merchantName": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "amount2": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency2": {
            "type": [
              "null",
              "string"
            ]
          },
          "items": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CreateTransactionItemRequest"
            }
          }
        }
      },
      "UpdateTransactionResponse": {
        "required": [
          "id",
          "date",
          "type",
          "amount",
          "amount2",
          "currency",
          "currency2",
          "accountId",
          "tripId",
          "categoryId",
          "subCategoryId",
          "targetAccountId",
          "originalTransactionId",
          "isRefund",
          "note",
          "merchantName",
          "location",
          "items"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "amount2": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency2": {
            "type": [
              "null",
              "string"
            ]
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subCategoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetAccountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "originalTransactionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isRefund": {
            "type": "boolean"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "merchantName": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionItemResponse"
            }
          }
        }
      },
...
    }
  },
...
}